mooved navigation component to pages folder
