+++ packages/pn-commons/src/index.ts
@@ -2,9 +2,10 @@ export * from './api';
 export * from './components';
 export * from './hooks';
 export * from './models';
-export * from './navigation';
+export * from './pages';
 export * from './redux';
 export * from './utility';
 export * from './validators';
+export * from './services';
 // TODO: questo export crea problemi nei test di pf e pg. Capire perchè e indagare il codice in setupTests
 // export { testSelect, testAutocomplete } from './test-utils';
