+++ packages/pn-commons/src/pages/index.ts
@@ -1,0 +1,5 @@
+import AccessDenied from './AccessDenied';
+import CourtesyPage from './CourtesyPage';
+import NotFound from './NotFound';
+
+export { NotFound, AccessDenied, CourtesyPage };
