+++ packages/pn-commons/src/services/configuration.service.ts
@@ -3,7 +3,7 @@
 /* eslint-disable functional/immutable-data */
 import { Validator } from '@pagopa-pn/pn-validator';
 
-import { fetchConfiguration } from './fetch.configuration.utility';
+import { fetchConfiguration } from '../utility/fetch.configuration.utility';
 
 class ConfigurationError extends Error {}
