+++ packages/pn-commons/src/services/index.ts
@@ -1,0 +1,4 @@
+import { Configuration } from './configuration.service';
+import { interceptDispatch, trackEvent } from './tracking.service';
+
+export { trackEvent, interceptDispatch, Configuration };
